ml-service/requirements.txt
fastapi
uvicorn
librosa
numpy
scikit-learn
soundfile
python-multipart
openai-whisper
websockets
torch
